requirements.txt
beautifulsoup4
gspread
google-auth
pandas
python-dateutil
PyYAML
requests
tenacity
